// src/app/shared/models/Devis.ts
import { Ticket } from "./Ticket";
import { DevisLine } from "./DevisLine";
import { Distributor } from "./Distributor";
import { Lead } from "./Lead";
import { Commercial } from "./Commercial";
import { StatusDevis } from "./StatusDevis";
import { Currency } from "./Currency";
import { Account } from "./accounts";
import { AddressInfo } from "./AddressInfo";
import { OrderType } from "./order-type";
import { OrderStatus } from "./OrderStatus";
import { Owner } from "./Owner";
import { User } from "./user";
export class Devis {
  id: number;
  owner: Owner;
  code: string;
  updateDate: Date;
  creationDate: Date;
  description: string;
  account: Account;
  orderStatus: OrderStatus;
  orderType: OrderType;
  remarks: string;
  lines: DevisLine[];
  addressByDeliveryDeliveryAddress: AddressInfo;
  cmdDeliveryNumberLine: number;
  cmdDeliveryNumberLineCmd: number;
  discount: number;
  user: User;
  commercial: Commercial;
  totalPriceHT: number;
  totalPriceTTC: number;
  version: number;
  currency: Currency;
  vat: number;
  expectedDate: Date;
  active: boolean;
  lead: Lead;
  statusDevis: StatusDevis;
  ticket: Ticket;
  saleOrder: string;
  pathSaleOrder: string;
  distributor: Distributor;
}